+++ inc/admin/wa-metaboxes.php
@@ -1,32 +1,30 @@
 <?php
 /**
- * %NAME% register metaboxes
+ * Wolf Albums register metaboxes
  *
- * @author %AUTHOR%
+ * @author WolfThemes
  * @category Core
- * @package %PACKAGENAME%/Admin
- * @version %VERSION%
+ * @package WolfAlbums/Admin
+ * @version 1.0.7
  */
 
-if ( ! defined( 'ABSPATH' ) ) {
-	exit;
-}
+defined( 'ABSPATH' ) || exit;
 
 $metaboxes = array(
 	'Album option' => array(
-		'title' => esc_html__( 'Album option', '%TEXTDOMAIN%' ),
+		'title' => esc_html__( 'Album option', 'wolf-albums' ),
 		'page' => array( 'gallery' ),
 		'metafields' => array(
 			array(
-				'label'	=> esc_html__( 'Cover Image Size', '%TEXTDOMAIN%' ),
+				'label'	=> esc_html__( 'Cover Image Size', 'wolf-albums' ),
 				'id'	=> '_wolf_album_image_size',
 				'type'	=> 'select',
 				'options' => array(
-					'wpb-2x1' => esc_html__( 'Standard', '%TEXTDOMAIN%' ),
-					'wpb-1x1' => esc_html__( 'Small Square', '%TEXTDOMAIN%' ),
-					'wpb-1x2' => esc_html__( 'Portrait', '%TEXTDOMAIN%' ),
-					'wpb-2x1' => esc_html__( 'Landscape', '%TEXTDOMAIN%' ),
-					'wpb-2x2' => esc_html__( 'Big Square', '%TEXTDOMAIN%' ),
+					'wpb-2x1' => esc_html__( 'Standard', 'wolf-albums' ),
+					'wpb-1x1' => esc_html__( 'Small Square', 'wolf-albums' ),
+					'wpb-1x2' => esc_html__( 'Portrait', 'wolf-albums' ),
+					'wpb-2x1' => esc_html__( 'Landscape', 'wolf-albums' ),
+					'wpb-2x2' => esc_html__( 'Big Square', 'wolf-albums' ),
 				),
 			),
 		),