+++ inc/frontend/wa-template-functions.php
@@ -1,18 +1,16 @@
 <?php
 /**
- * %NAME% Template Functions
+ * Wolf Albums Template Functions
  *
  * Functions used in the template files to output content - in most cases hooked in via the template actions. All functions are pluggable.
  *
- * @author %AUTHOR%
+ * @author WolfThemes
  * @category Core
- * @package %PACKAGENAME%/Templates
+ * @package WolfAlbums/Templates
  * @since 1.2.2
  */
 
-if ( ! defined( 'ABSPATH' ) ) {
-	exit; // Exit if accessed directly
-}
+defined( 'ABSPATH' ) || exit;
 
 /**
  * Output generator tag to aid debugging.